Version 0.2.1
This is basically applying a version number to the changes already made.

I am following the logic A.B.C where A is for complete changes to the
plugin code or to the way it works; B is for increased functionality
(i.e. the plugin does new things); and C for all other changes which do
not add functionality.

I’m merging readme.txt into readme.md to avoid duplication in
documentation.